Move header-only c10/util low-level utils to torch/headeronly
Migrate five already-header-only c10/util headers into torch/headeronly/util/,
leaving the old c10 paths as one-line forwarders (fully backward compatible, no
consumer include changes):

  irange.h, Load.h, Unroll.h, python_stub.h, static_tracepoint_elfx86.h

Part of the wave moving c10 headers that need no .cpp/libtorch linking into the
header-only library. irange/Load/Unroll are flipped (defined canonically in
torch::headeronly, re-aliased into c10). python_stub.h (a global PyObject
forward declaration) and static_tracepoint_elfx86.h (pure preprocessor macros)
have no namespaced symbols, so they are plain moves. irange keeps its existing
NOLINT(modernize-use-constraints) intent on the SFINAE overloads.

Test Plan:

Full build (confirms all libtorch consumers still compile):

```
pip install -e . --no-build-isolation
```

Header-only ABI check (compiles + runs without linking libtorch):

```
build/bin/test_aoti_abi_check --gtest_filter='TestIrange*:TestLoad*:TestUnroll*'
```

Lint:

```
lintrunner -a <changed files>
```

This PR was authored with the assistance of Claude (Claude Code).

[ghstack-poisoned]
