+++ torch/headeronly/util/python_stub.h
@@ -1,0 +1,4 @@
+#pragma once
+
+struct _object;
+using PyObject = _object;
